+++ modules/libcom/src/as/asTrapWrite.h
@@ -36,6 +36,7 @@ typedef struct asTrapWriteMessage {
     const char *method; /**< \brief Method of the request (ca, anonymous, or x509. */
     const char *authority; /**< \brief Authority of the request (common name of the root certificate authority). */
     const char *hostid; /**< \brief Hostid of whoever originated the request. */
+    int isTLS; /**< \brief the connection is TLS encapsulated. */
     /** \brief A field for use by the server.
      *
      * Any listener that uses this field must know what type of